Editar obra "terminado"

+++ src/main/java/com/registro/obras/Modelo/ObraMantencion.java
@@ -28,7 +28,7 @@ public class ObraMantencion extends Obra{
         return InteresAnual;
     }
 
-    public void setInteresAnual(long InteresAnual) {
+    public void setInteresAnual(double InteresAnual) {
         this.InteresAnual = InteresAnual;
     }
